meiduo_mall/requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.11
django-ckeditor==5.5.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.0.1
django-js-asset==1.1.0
django-redis==4.9.0
django-reversion==2.0.13
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
drf-haystack==1.8.3
elasticsearch==2.4.1
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.16.0
httplib2==0.9.2
idna==2.7
itsdangerous==0.24
jdcal==1.4
kombu==4.2.1
mutagen==1.40.0
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.1.0
pkg-resources==0.0.0
pycryptodomex==3.6.3
PyJWT==1.6.4
PyMySQL==0.8.1
python-alipay-sdk==1.7.1
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
redis==2.10.6
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
vine==1.1.4
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
